vibrator: Implement Effect::TICK
The AOSP TICK effect generated from the following test app:
- https://github.com/luk1337/VibeTest
was used as the starting point. It is more or less the same as oplus
CLICK, except waveform_index is set to 2.

The vibration effect from this TICK was stronger than oplus CLICK, which
is not really desirable, so adjustments were made to vmax in an
unscientific manner until it felt and sounded just right.

Change-Id: I472db022c2b0182e9a0d544c9d439bc3e47e4d82
Signed-off-by: Daryl Ronningen <relms@relms.dev>